// client/src/app/modules/core/types/stationsInterface.ts
export interface StationsInterface {
  stationId: number;
  scopeOfApproval: string;
  postalCode: string;
  city: string;
  street: string;
  operator: string;
  telephoneNumber: string;
  email: string;
  community: string;
  district: string;
  region: string;
  inspectionPrice: number;
  evInspectionPrice: number;
  emissionPriceB: number;
  emissionPriceN: number;
  workingHours: string;
}